+++ client/src/lib/session/components/Slides/Slides/Blocks/Video.tsx
@@ -7,10 +7,10 @@ import DurationTimer from '../../../DurationTimer/DurationTimer';
 import {LottiePlayerHandle} from '../../../../../components/LottiePlayer/LottiePlayer';
 import {VideoLooperProperties} from '../../../../../../../types/VideoLooper';
 import VideoLooper from '../../../../../components/VideoLooper/VideoLooper';
-import MediaControls from '../../../MediaControls/MediaControls';
+import MediaControls from '../../../../../components/MediaControls/MediaControls';
 import {Spacer16} from '../../../../../components/Spacers/Spacer';
 import MediaWrapperResolver from './MediaWrapperResolver';
-import Subtitles from './Subtitles';
+import Subtitles from '../../../../../components/Subtitles/Subtitles';
 import Gutters from '../../../../../components/Gutters/Gutters';
 
 const VideoPlayer = styled(VideoLooper)({
